requirements.txt
flask_sqlalchemy==2.4.1
Flask-Cors==3.0.8
flask_marshmallow==0.11.0
marshmallow-sqlalchemy==0.22.2
